requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
httpx==0.24.0
pydantic==1.10.7
supabase==1.0.3
redis==4.5.4
clerk-sdk-python==0.1.1
anthropic==0.5.0
pytest==7.3.1
pytest-asyncio==0.21.0
